Analytical Study On Training And Assessment Practices Adopted By Employees Of Haldirams Pvt Ltd. Nagpur
One of the most talked subjects in corporate circles, in recent times is how to optimize the contributions of human resources in achieving organizational goals. An efficient and satisfied workforce is the most significant factor in organizational effectiveness and marginal excellence. But experiences in business and service organizations however indicate that management, comparatively speaking, best more time and attention to policies and systems relating to production, technology, investment, inventory, marketing, etc than to human resources. Compulsion of modern business environment is however promoting corporate managements to systematically review their current attitudes, beliefs and policies towards human resources for they contribute most significantly to the survival and growth of organization. Enlightened and progressive managements have therefore come to believe that people are not problems but opportunities. Training and Assessment activities are important part of exploring these opportunities. Training and Assessment activities are designed to impart specific skills, abilities and knowledge to employees. Training is a planned effort by an organization to facilitate employee’s learning of job related knowledge and skills for the purpose of improving performance. Assessment refers to learning opportunities designed to help employees grow and evolve a vision for future.